CMCP-05: Separate MCP server from REST API
- Split mcp_app.py into mcp_server.py (MCP tools only) and rest_api.py (REST endpoints)
- Updated asgi.py to mount both apps at /mcp/ and / respectively
- mcp_app.py now minimal (11 lines) with backward compatibility imports
- All CORS handling preserved in rest_api.py
- Server architecture now properly separated
